mobile bug fixes and responsive bottom mobile menu, pwa icons and title and color added

+++ src/app/software/company/list-company/list-company.component.ts
@@ -45,9 +45,13 @@ export class ListCompanyComponent implements OnInit {
   }
 
   public onDeleteCompany(idToDelete: number): void {
-    this.companyService.deleteCompany({ databaseId: idToDelete }).subscribe(res => {
-      this.utility.message('شرکت با موفقیت حذف شد.', 'بستن');
-      this.loadCompanyList()
+    this.dialog.openAcceptDeleteDialog().afterClosed().subscribe(result => {
+      if (result) {
+        this.companyService.deleteCompany({ databaseId: idToDelete }).subscribe(res => {
+          this.utility.message('شرکت با موفقیت حذف شد.', 'بستن');
+          this.loadCompanyList();
+        })
+      }
     })
   }
